app/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
djangorestframework==3.12.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
kombu==5.0.2
oauthlib==3.1.0
packaging==20.9
prompt-toolkit==3.0.16
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
gunicorn==20.1.0

